requirements.txt
anyio==3.6.2
asgiref==3.5.2
black==22.10.0
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dj-database-url==1.0.0
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fastapi==0.85.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
mypy-extensions==0.4.3
numpy==1.23.4
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.1
pandas-stubs==1.5.1.221024
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.20.4
tomli==2.0.1
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
whitenoise==6.2.0
